cpp modularization
