+++ src/pages/Notes/JavaScript/sections/IntroSection.jsx
@@ -1,60 +1,35 @@
 import React from "react";
+import { FaCopy } from "react-icons/fa";
 
 const IntroSection = ({ copyCode, copiedCode }) => {
+  const helloWorldCode = `console.log("Hello, World!");`;
+
   return (
     <section style={{ marginBottom: "2rem" }}>
       <div className="card">
-        <h2><i className="fas fa-play-circle"></i> 1. Introduction to JavaScript</h2>
-        <p>JavaScript is a versatile, high-level programming language that powers the modern web. Originally created for web browsers, it's now used for servers, mobile apps, desktop applications, and more.</p>
-        
+        <h2>
+          <i className="fab fa-js-square" style={{ marginRight: "0.5rem" }}></i>
+          Introduction to JavaScript
+        </h2>
+        <p>
+          JavaScript is a high-level, interpreted programming language that is
+          one of the core technologies of the World Wide Web, alongside HTML and
+          CSS. It enables interactive web pages and is an essential part of web
+          applications.
+        </p>
         <h3>Key Features</h3>
         <ul>
-          <li><strong>Dynamic:</strong> Variables can change types at runtime</li>
-          <li><strong>Interpreted:</strong> No compilation step needed</li>
-          <li><strong>Event-driven:</strong> Responds to user interactions</li>
-          <li><strong>Cross-platform:</strong> Runs everywhere - browsers, servers, mobile</li>
-          <li><strong>Object-oriented:</strong> Supports objects and classes</li>
+          <li><strong>Client-Side Scripting:</strong> Runs directly in the user's browser.</li>
+          <li><strong>Asynchronous:</strong> Handles operations like network requests without blocking the main thread.</li>
+          <li><strong>Dynamically Typed:</strong> Variable types are determined at runtime.</li>
+          <li><strong>Vast Ecosystem:</strong> Huge number of libraries and frameworks (React, Angular, Vue.js).</li>
         </ul>
-
-        <h3>Your First JavaScript Program</h3>
-        <p>Let's start with a simple "Hello, World!" program:</p>
         <div className="code-container">
-          <button 
-            className={`copy-btn ${copiedCode === "intro" ? "copied" : ""}`} 
-            onClick={() => copyCode(`// Your first JavaScript program
-console.log("Hello, World!");
-
-// Variables and basic operations
-let name = "JavaScript";
-let year = 2024;
-let isAwesome = true;
-
-console.log("Welcome to " + name + "!");
-console.log("Current year: " + year);
-console.log("Is JavaScript awesome? " + isAwesome);`, "intro")}
-          >
-            {copiedCode === "intro" ? "Copied!" : "Copy"}
+          <pre><code>{helloWorldCode}</code></pre>
+          <button className={`copy-btn ${copiedCode === "intro" ? "copied" : ""}`} onClick={() => copyCode(helloWorldCode, "intro")}>
+            <FaCopy />
           </button>
-          <pre>{`// Your first JavaScript program
-console.log("Hello, World!");
-
-// Variables and basic operations
-let name = "JavaScript";
-let year = 2024;
-let isAwesome = true;
-
-console.log("Welcome to " + name + "!");
-console.log("Current year: " + year);
-console.log("Is JavaScript awesome? " + isAwesome);`}</pre>
         </div>
-
-        <h3>Where JavaScript Runs</h3>
-        <ul>
-          <li><strong>Web Browsers:</strong> Chrome, Firefox, Safari, Edge</li>
-          <li><strong>Servers:</strong> Node.js for backend development</li>
-          <li><strong>Mobile Apps:</strong> React Native, Ionic</li>
-          <li><strong>Desktop Apps:</strong> Electron (VS Code, Discord)</li>
-        </ul>
       </div>
     </section>
   );
